refactor: reorganize HostMainPage component and update import paths
